backend/requirements.txt
# Core dependencies
insightface>=0.7.3
onnxruntime>=1.16.0
opencv-python>=4.8.0
rawpy>=0.19.0
Pillow>=10.0.0
watchdog>=3.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Google Drive API
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
